# api/service/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: carl.a.smigielski@saic.com
    name: Carl Smigielski
  description: An API for managing evaluations of Security Technical Implementation
    Guide (STIG) assessments.
  title: STIG Management API
  version: 1.0.0
servers:
- url: http://localhost:54000/stig-manager/api
paths:
  /packages:
    get:
      operationId: getPackages
      parameters:
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - assets
            - stigs
            type: string
          type: array
        style: form
      - description: Elevate the user context for this request if user is permitted
          (canAdmin)
        explode: true
        in: query
        name: elevate
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PackageProjected'
                type: array
          description: PackageProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package:read
      summary: Return a list of Packages accessible to the requester
      tags:
      - Package
      x-classic:
      - getPackages
      - getPackageProps
      x-swagger-router-controller: Package
    post:
      operationId: createPackage
      parameters:
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - assets
            - stigs
            type: string
          type: array
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageAssign'
        required: true
        x-name: body
      responses:
        "201":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PackageProjected'
                type: array
          description: PackageProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package
      summary: Create a Package
      tags:
      - Package
      x-codegen-request-body-name: body
      x-swagger-router-controller: Package
  /packages/{packageId}:
    delete:
      operationId: deletePackage
      parameters:
      - description: A path parameter that indentifies a Package
        explode: false
        in: path
        name: packageId
        required: true
        schema:
          type: integer
        style: simple
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - assets
            - stigs
            type: string
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageProjected'
          description: PackageProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package
      summary: Delete a Package
      tags:
      - Package
      x-swagger-router-controller: Package
    get:
      operationId: getPackage
      parameters:
      - description: A path parameter that indentifies a Package
        explode: false
        in: path
        name: packageId
        required: true
        schema:
          type: integer
        style: simple
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - assets
            - stigs
            type: string
          type: array
        style: form
      - description: Elevate the user context for this request if user is permitted
          (canAdmin)
        explode: true
        in: query
        name: elevate
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageProjected'
          description: PackageProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package:read
      summary: Return a Package
      tags:
      - Package
      x-swagger-router-controller: Package
    patch:
      operationId: updatePackage
      parameters:
      - description: A path parameter that indentifies a Package
        explode: false
        in: path
        name: packageId
        required: true
        schema:
          type: integer
        style: simple
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - assets
            - stigs
            type: string
          type: array
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageAssign'
        required: true
        x-name: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageProjected'
          description: PackageProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package
      summary: Merge updates to a Package
      tags:
      - Package
      x-codegen-request-body-name: body
      x-swagger-router-controller: Package
...
  /reviews:
    get:
      operationId: getReviews
      parameters:
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - asset
            - history
            - rule
            - stigs
            type: string
          type: array
        style: form
      - description: Elevate the user context for this request if user is permitted
          (canAdmin)
        explode: true
        in: query
        name: elevate
        required: false
        schema:
          type: boolean
        style: form
      - explode: true
        in: query
        name: state
        required: false
        schema:
          $ref: '#/components/schemas/ReviewState'
        style: form
      - explode: true
        in: query
        name: action
        required: false
        schema:
          $ref: '#/components/schemas/ReviewAction'
        style: form
      - explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: '#/components/schemas/ReviewStatus'
        style: form
      - description: Selects Reviews of a Rule
        explode: true
        in: query
        name: ruleId
        required: false
        schema:
          type: string
        style: form
      - description: Selects Reviews mapped to a STIG
        explode: true
        in: query
        name: benchmarkId
        required: false
        schema:
          type: string
        style: form
      - description: Selects Reviews mapped to an Asset
        explode: true
        in: query
        name: assetId
        required: false
        schema:
          type: string
        style: form
      - description: Selects Reviews mapped to a Package
        explode: true
        in: query
        name: packageId
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ReviewProjected'
                type: array
          description: ReviewProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package:read
      summary: Return a list of Reviews accessible to the requester
      tags:
      - Review
      x-swagger-router-controller: Review
    post:
      operationId: createReview
      parameters:
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - asset
            - history
            - rule
            - stigs
            type: string
          type: array
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
        x-name: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewProjected'
          description: ReviewProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package
      summary: Create a Review
      tags:
      - Review
      x-codegen-request-body-name: body
      x-swagger-router-controller: Review
  /reviews/{reviewId}:
    delete:
      operationId: deleteReview
      parameters:
      - description: A path parameter that indentifies a Review
        explode: false
        in: path
        name: reviewId
        required: true
        schema:
          type: integer
        style: simple
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - asset
            - history
            - rule
            - stigs
            type: string
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewProjected'
          description: ReviewProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package
      summary: Delete a Review
      tags:
      - Review
      x-swagger-router-controller: Review
    get:
      operationId: getReview
      parameters:
      - description: A path parameter that indentifies a Review
        explode: false
        in: path
        name: reviewId
        required: true
        schema:
          type: integer
        style: simple
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - asset
            - history
            - rule
            - stigs
            type: string
          type: array
        style: form
      - description: Elevate the user context for this request if user is permitted
          (canAdmin)
        explode: true
        in: query
        name: elevate
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetProjected'
          description: ReviewProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package:read
      summary: Return a Review
      tags:
      - Review
      x-swagger-router-controller: Review
    patch:
      operationId: updateReview
      parameters:
      - description: A path parameter that indentifies a Review
        explode: false
        in: path
        name: reviewId
        required: true
        schema:
          type: integer
        style: simple
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - asset
            - history
            - rule
            - stigs
            type: string
          type: array
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetAssign'
        required: true
        x-name: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewProjected'
          description: ReviewProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:package
      summary: Merge updates to a Review
      tags:
      - Review
      x-codegen-request-body-name: body
      x-swagger-router-controller: Review
...
components:
...
  schemas:
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object
    STIG:
      example:
        lastRevisionDate: lastRevisionDate
        title: title
        benchmarkId: benchmarkId
        lastRevisionStr: lastRevisionStr
      properties:
        benchmarkId:
          type: string
        title:
          type: string
        lastRevisionStr:
          type: string
        lastRevisionDate:
          type: string
      required:
      - benchmarkId
      type: object
...
    RuleAbbr:
      properties:
        ruleId:
          type: string
        version:
          type: string
        title:
          type: string
        severity:
          type: string
      type: object
...
    MacCl:
      enum:
      - MAC-1_Classified
      - MAC-1_Sensitive
      - MAC-1_Public
      - MAC-2_Classified
      - MAC-2_Sensitive
      - MAC-2_Public
      - MAC-3_Classified
      - MAC-3_Sensitive
      - MAC-3_Public
      type: string
    Package:
      properties:
        packageId:
          nullable: false
          readOnly: true
          type: integer
        name:
          nullable: false
          type: string
        emassId:
          nullable: true
          type: string
        pocName:
          nullable: true
          type: string
        pocEmail:
          nullable: true
          type: string
        pocPhone:
          nullable: true
          type: string
        reqRar:
          nullable: true
          type: boolean
        macCl:
          $ref: '#/components/schemas/MacCl'
      required:
      - name
      - packageId
      type: object
    PackageBasic:
      properties:
        packageId:
          type: integer
        name:
          type: string
      type: object
    PackageProjected:
      allOf:
      - $ref: '#/components/schemas/Package'
      - $ref: '#/components/schemas/PackageProjected_allOf'
    PackageAssign:
      allOf:
      - $ref: '#/components/schemas/Package'
      - $ref: '#/components/schemas/PackageAssign_allOf'
    Asset:
      properties:
        assetId:
          readOnly: true
          type: integer
        name:
          nullable: false
          type: string
        dept:
          nullable: false
          type: string
        ip:
          nullable: true
          type: string
        nonnetwork:
          nullable: true
          type: boolean
        scanexempt:
          nullable: true
          type: boolean
        profile:
          $ref: '#/components/schemas/MacCl'
      required:
      - assetId
      - dept
      - name
      type: object
    AssetProjected:
      allOf:
      - $ref: '#/components/schemas/Asset'
      - $ref: '#/components/schemas/AssetProjected_allOf'
    AssetBasic:
      properties:
        assetId:
          type: integer
        name:
          type: string
        dept:
          type: string
      type: object
    AssetAssign:
      allOf:
      - $ref: '#/components/schemas/Asset'
      - $ref: '#/components/schemas/AssetAssign_allOf'
    StigAssigned:
      properties:
        saId:
          type: integer
        benchmarkId:
          type: string
      type: object
...
    ReviewState:
      description: The Review state
      enum:
      - "0"
      - NF
      - NA
      type: string
    ReviewAction:
      enum:
      - Remediate
      - Mitigate
      - Exception
      type: string
    ReviewStatus:
      enum:
      - Saved
      - Submitted
      - Approved
      - Rejected
      type: string
    Review:
      example:
        rejectText: rejectText
        actionComment: actionComment
        userId: 1
        rejectUserId: 5
        assetId: 6
        reqDoc: true
        stateComment: stateComment
        ruleId: ruleId
        reviewId: 0
        autoState: true
        ts: ts
      properties:
        reviewId:
          readOnly: true
          type: integer
        assetId:
          type: integer
        ruleId:
          type: string
        state:
          $ref: '#/components/schemas/ReviewState'
        stateComment:
          type: string
        autoState:
          type: boolean
        action:
          $ref: '#/components/schemas/ReviewAction'
        actionComment:
          type: string
        reqDoc:
          type: boolean
        status:
          $ref: '#/components/schemas/ReviewStatus'
        userId:
          readOnly: true
          type: integer
        ts:
          readOnly: true
          type: string
        rejectText:
          type: string
        rejectUserId:
          type: integer
      type: object
    ReviewProjected:
      allOf:
      - $ref: '#/components/schemas/Review'
      - $ref: '#/components/schemas/ReviewProjected_allOf'
...
    PackageProjected_allOf:
      properties:
        assets:
          items:
            $ref: '#/components/schemas/AssetBasic'
          type: array
        stigs:
          items:
            $ref: '#/components/schemas/STIG'
          type: array
    PackageAssign_allOf:
      properties:
        assetIds:
          items:
            type: integer
          type: array
    AssetProjected_allOf:
      properties:
        packages:
          description: The Packages mapped to this Asset
          items:
            $ref: '#/components/schemas/PackageBasic'
          type: array
        stigs:
          description: The benchmarkIds of STIGs mapped to this Asset
          items:
            $ref: '#/components/schemas/StigAssigned'
          type: array
    AssetAssign_allOf:
      properties:
        packageIds:
          description: The packageIds mapped to this Asset
          items:
            type: integer
          type: array
        benchmarkIds:
          description: The benchmarkIds of STIGs mapped to this Asset
          items:
            type: string
          type: array
...
    ReviewProjected_allOf:
      properties:
        history:
          items:
            type: string
          type: array
        asset:
          $ref: '#/components/schemas/AssetBasic'
        stigs:
          items:
            type: string
          type: array
        rule:
          $ref: '#/components/schemas/RuleAbbr'
  securitySchemes:
    oauth:
      flows:
        implicit:
          authorizationUrl: http://localhost:8080/auth/realms/cmsat/protocol/openid-connect/auth
          scopes:
            stig-manager: Top-level scope
            stig-manager:stig: Full access to the STIG collection
            stig-manager:stig:read: Read access to the STIG collection
            stig-manager:package: Full access to the Package collection
            stig-manager:package:read: Read-only access to the Package collection
            stig-manager:user: Full access to the Users collection
            stig-manager:user:read: Read-only access to the Users collection
      type: oauth2